Handle mongod messages and let mongod find a free port

+++ main.go
@@ -4,18 +4,21 @@ import (
 	"bufio"
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"net"
 	"os"
 	"os/exec"
-	"strconv"
+	"time"
 
 	"github.com/ONSdigital/dp-mongodb-in-memory/download"
 	"github.com/ONSdigital/dp-mongodb-in-memory/monitor"
 	"github.com/ONSdigital/log.go/v2/log"
 )
+
+// max time allowed for mongo to start
+const timeout = 5 * time.Second
 
 // Server represents a running MongoDB server.
 type Server struct {
@@ -32,11 +35,7 @@ func init() {
 // Start runs a MongoDB server at a given version using a random free port
 // and returns the Server.
 func Start(version string) (*Server, error) {
-	port, err := getFreePort()
-	if err != nil {
-		log.Fatal(context.Background(), "Could not find a free port", err)
-		return nil, err
-	}
+	server := new(Server)
 
 	binPath, err := getOrDownloadBinPath(version)
 	if err != nil {
@@ -45,70 +44,71 @@ func Start(version string) (*Server, error) {
 	}
 
 	// Create a db dir. Even the ephemeralForTest engine needs a dbpath.
-	dbDir, err := ioutil.TempDir("", "")
+	server.dbDir, err = ioutil.TempDir("", "")
 	if err != nil {
 		log.Fatal(context.Background(), "Error creating data directory", err)
 		return nil, err
 	}
 
-	log.Info(context.Background(), "Starting mongod server", log.Data{"binPath": binPath, "dbDir": dbDir, "port": port})
-	cmd := exec.Command(binPath, "--storageEngine", "ephemeralForTest", "--dbpath", dbDir, "--port", strconv.Itoa(port))
+	log.Info(context.Background(), "Starting mongod server", log.Data{"binPath": binPath, "dbDir": server.dbDir})
+	server.cmd = exec.Command(binPath, "--storageEngine", "ephemeralForTest", "--dbpath", server.dbDir, "--port", "0")
 
-	cmd.Stderr = stdHandler()
-	cmd.Stdout = cmd.Stderr
+	startupErrCh := make(chan error)
+	startupPortCh := make(chan int)
+	stdHandler := stdHandler(startupPortCh, startupErrCh)
+	server.cmd.Stdout = stdHandler
+	server.cmd.Stderr = stdHandler
 
 	// Run the server
-	err = cmd.Start()
+	err = server.cmd.Start()
 	if err != nil {
-		remErr := os.RemoveAll(dbDir)
-		if remErr != nil {
-			log.Error(context.Background(), "Error removing data directory", remErr, log.Data{"dir": dbDir})
-		}
 		log.Fatal(context.Background(), "Could not start mongodb", err)
+		server.Stop()
 		return nil, err
 	}
 
 	log.Info(context.Background(), "Starting watcher")
 	// Start a watcher: the watcher is a subprocess that ensures if this process
 	// dies, the mongo server will be killed (and not reparented under init)
-	watcherCmd, err := monitor.Run(os.Getpid(), cmd.Process.Pid)
+	server.watcherCmd, err = monitor.Run(os.Getpid(), server.cmd.Process.Pid)
 	if err != nil {
 		log.Error(context.Background(), "Could not start watcher", err)
-
-		killErr := cmd.Process.Kill()
-		if killErr != nil {
-			log.Error(context.Background(), "Error stopping mongod process", killErr)
-		}
-
-		remErr := os.RemoveAll(dbDir)
-		if remErr != nil {
-			log.Error(context.Background(), "Error removing data directory", err, log.Data{"dir": dbDir})
-		}
-
+		server.Stop()
 		return nil, err
 	}
 
-	return &Server{
-		cmd:        cmd,
-		watcherCmd: watcherCmd,
-		dbDir:      dbDir,
-		port:       port,
-	}, nil
+	select {
+	case server.port = <-startupPortCh:
+	case err := <-startupErrCh:
+		server.Stop()
+		return nil, err
+	case <-time.After(timeout):
+		server.Stop()
+		return nil, errors.New("timed out waiting for mongod to start")
+	}
+
+	log.Info(context.Background(), fmt.Sprintf("mongod started up and reported port number %d", server.port))
+
+	return server, nil
 }
 
 // Stop kills the mongo server.
 func (s *Server) Stop() {
-	err := s.cmd.Process.Kill()
-	if err != nil {
-		log.Error(context.Background(), "Error stopping mongod process", err, log.Data{"pid": s.cmd.Process.Pid})
+	if s.cmd != nil {
+		err := s.cmd.Process.Kill()
+		if err != nil {
+			log.Error(context.Background(), "Error stopping mongod process", err, log.Data{"pid": s.cmd.Process.Pid})
+		}
 	}
 
-	err = s.watcherCmd.Process.Kill()
-	if err != nil {
-		log.Error(context.Background(), "error stopping watcher process", err, log.Data{"pid": s.watcherCmd.Process.Pid})
+	if s.watcherCmd != nil {
+		err := s.watcherCmd.Process.Kill()
+		if err != nil {
+			log.Error(context.Background(), "error stopping watcher process", err, log.Data{"pid": s.watcherCmd.Process.Pid})
+		}
 	}
 
-	err = os.RemoveAll(s.dbDir)
+	err := os.RemoveAll(s.dbDir)
 	if err != nil {
 		log.Error(context.Background(), "Error removing data directory", err, log.Data{"dir": s.dbDir})
 	}
@@ -137,32 +137,32 @@ func getOrDownloadBinPath(version string) (string, error) {
 	return config.MongoPath(), nil
 }
 
-func getFreePort() (int, error) {
-	// Based on: https://github.com/phayes/freeport/blob/master/freeport.go
-	addr, err := net.ResolveTCPAddr("tcp", "localhost:0")
-	if err != nil {
-		return 0, err
-	}
-
-	l, err := net.ListenTCP("tcp", addr)
-	if err != nil {
-		return 0, err
-	}
-	defer l.Close()
-	return l.Addr().(*net.TCPAddr).Port, nil
-}
-
-// stdHandler handler just relays messages from stdout/stderr to our logger
-func stdHandler() io.Writer {
+// stdHandler handler relays messages from stdout/stderr to our logger.
+// It accepts 2 channels:
+// errCh will receive any error logged,
+// okCh will receive the port number if mongodb started successfully
+func stdHandler(okCh chan<- int, errCh chan<- error) io.Writer {
 	reader, writer := io.Pipe()
 
 	go func() {
 		scanner := bufio.NewScanner(reader)
 
 		for scanner.Scan() {
-			var jsonError log.Data
-			json.Unmarshal([]byte(scanner.Text()), &jsonError)
-			log.Info(context.Background(), fmt.Sprintf("[mongod] %s", jsonError["msg"]), jsonError)
+			var logMessage log.Data
+			json.Unmarshal([]byte(scanner.Text()), &logMessage)
+
+			message := logMessage["msg"]
+			severity := logMessage["s"]
+			if severity == "E" || severity == "F" {
+				// error or fatal
+				errCh <- errors.New(fmt.Sprintf("Mongod startup failed: %s", message))
+			} else if severity == "I" && message == "Waiting for connections" {
+				// Mongo running successfully: find port
+				attr := logMessage["attr"].(map[string]interface{})
+				okCh <- int(attr["port"].(float64))
+			}
+
+			log.Info(context.Background(), fmt.Sprintf("[mongod] %s", message), logMessage)
 		}
 
 		if err := scanner.Err(); err != nil {
